move files back, print informatino
