+++ scripts/gcp_setup/environment/utils.py
